Fixes #193 Refactored mapcomponents.js and associated helper functions to cluster the 311 data according to both complaint type and severity. Removed 1 park from outside the manhattan polygon
