+++ frontend/src/components/MapComponent/MapHelper/markerHelpers.js
@@ -1,5 +1,8 @@
 import mapboxgl from 'mapbox-gl';
 import $ from 'jquery';
+import ReactDOM from 'react-dom';
+import PopupContent from '../../PopupContent/PopupContent';
+
 
 export function addMarkers(mapRef, markerData, markerType) {
   markerData.forEach(location => {
@@ -21,7 +24,7 @@ export function add311Markers(mapRef, markerData, category) {
   markerData.forEach(location => {
     const marker = document.createElement('div');
     const markerType = getMarkerType(category, location.category);
-    marker.className = `marker ${markerType}_marker`;
+    marker.className = `marker ${markerType}_marker marker311`;
 
     new mapboxgl.Marker({
       element: marker,
@@ -33,6 +36,15 @@ export function add311Markers(mapRef, markerData, category) {
 
     animateMarkers($(marker));
   });
+}
+function animateMarkers($marker) {
+  $marker.css({
+    top: '-50px',
+    opacity: 0,
+  }).animate({
+    top: '0px',
+    opacity: 1,
+  }, 500);
 }
 
 // Function to add 311 multiple complaint markers
@@ -50,9 +62,12 @@ export function add311Multiple(mapRef, markerData) {
       .setPopup(new mapboxgl.Popup({ offset: 25 }).setText(location.Complaint_Type_collection))
       .addTo(mapRef.current);
 
-    animateMarkers($(marker));
-  });
+    animateMarkers($(marker));}
+  );
 }
+
+
+
 function getMarkerType(complaintCategory, severityCategory) {
   let markerType = complaintCategory.toLowerCase();
 
@@ -88,17 +103,24 @@ export function plotRoutePOI(mapRef, geojsonData, bounds) {
   };
 
   // Add the filtered POIs as a clustered source
-  addClusteredLayer(mapRef, filteredGeojsonData, 'route-pois', 'poi-marker');
+  addClusteredLayer(mapRef, filteredGeojsonData, 'route-pois', 'poi-marker', '#00cffa');
 }
 
 
-export function addClusteredLayer (mapRef, geoJsonData, sourceId, imageId) {
+export function addClusteredLayer(mapRef, geoJsonData, sourceId, imageId, clusterColor) {
+  if (mapRef.current.getSource(sourceId)) {
+    mapRef.current.removeLayer(`${sourceId}-clusters`);
+    mapRef.current.removeLayer(`${sourceId}-cluster-count`);
+    mapRef.current.removeLayer(`${sourceId}-unclustered-point`);
+    mapRef.current.removeSource(sourceId);
+  }
+
   mapRef.current.addSource(sourceId, {
     type: 'geojson',
     data: geoJsonData,
     cluster: true,
     clusterMaxZoom: 14,
-    clusterRadius: 50
+    clusterRadius: 80
   });
 
   // Clusters are displayed using circles
@@ -108,15 +130,7 @@ export function addClusteredLayer (mapRef, geoJsonData, sourceId, imageId) {
     source: sourceId,
     filter: ['has', 'point_count'],
     paint: {
-      'circle-color': [
-        'step',
-        ['get', 'point_count'],
-        '#51bbd6',
-        100,
-        '#f1f075',
-        750,
-        '#f28cb1'
-      ],
+      'circle-color': clusterColor,
       'circle-radius': [
         'step',
         ['get', 'point_count'],
@@ -125,7 +139,8 @@ export function addClusteredLayer (mapRef, geoJsonData, sourceId, imageId) {
         30,
         750,
         40
-      ]
+      ],
+      'circle-opacity': 0.6
     }
   });
 
@@ -194,16 +209,210 @@ export function addClusteredLayer (mapRef, geoJsonData, sourceId, imageId) {
   mapRef.current.on('mouseleave', `${sourceId}-unclustered-point`, () => {
     mapRef.current.getCanvas().style.cursor = '';
   });
-};
+}
 
-function animateMarkers($marker) {
-  $marker.css({
-    top: '-50px',
-    opacity: 0,
-  }).animate({
-    top: '0px',
-    opacity: 1,
-  }, 500);
+
+export function clearMapFeatures(mapRef) {
+  // Function to ckear features close to the route
+  const sourceIds = ['route-pois','poi-marker','route-noise-h','route-noise-vh','route-garbage-h','route-other-h','route-multi-h','route-multi-vh']; 
+
+  sourceIds.forEach(sourceId => {
+    if (mapRef.current.getSource(sourceId)) {
+      mapRef.current.removeLayer(`${sourceId}-clusters`);
+      mapRef.current.removeLayer(`${sourceId}-cluster-count`);
+      mapRef.current.removeLayer(`${sourceId}-unclustered-point`);
+      mapRef.current.removeSource(sourceId);
+    }
+  });
+}
+
+export function addMapFeatures(mapRef, helpers) {
+  const {
+    fetchInitialPOI,
+    simulateFetchParks,
+    convertToGeoJSON,
+    floraImage,
+    noiseHighImage,
+    noiseVeryHighImage,
+    garbageHighImage,
+    otherHighImage,
+    multiHighImage,
+    multiVeryHighImage,
+    poiImage,
+    setStartCord,
+    setEndCord,
+    setWaypointAndIncrease,
+    updateStartInput,
+    updateEndInput,
+    updateWaypointInput,
+    geocoderRefs
+  } = helpers;
+
+  const initialPOI = fetchInitialPOI();
+  const parkData = simulateFetchParks();
+  addMarkers(mapRef, initialPOI, 'poi');
+  const parkGeoJson = convertToGeoJSON(parkData);
+
+  mapRef.current.loadImage(floraImage, (error, image) => {
+    if (error) throw error;
+    mapRef.current.addImage('flora-marker', image);
+  });
+
+  mapRef.current.loadImage(noiseHighImage, (error, image) => {
+    if (error) throw error;
+    mapRef.current.addImage('noise-high-marker', image);
+  });
+
+  mapRef.current.loadImage(noiseVeryHighImage, (error, image) => {
+    if (error) throw error;
+    mapRef.current.addImage('noise-veryhigh-marker', image);
+  });
+
+  mapRef.current.loadImage(garbageHighImage, (error, image) => {
+    if (error) throw error;
+    mapRef.current.addImage('garbage-high-marker', image);
+  });
+
+  mapRef.current.loadImage(otherHighImage, (error, image) => {
+    if (error) throw error;
+    mapRef.current.addImage('other-high-marker', image);
+  });
+
+  mapRef.current.loadImage(multiHighImage, (error, image) => {
+    if (error) throw error;
+    mapRef.current.addImage('multi-high-marker', image);
+  });
+
+  mapRef.current.loadImage(multiVeryHighImage, (error, image) => {
+    if (error) throw error;
+    mapRef.current.addImage('multi-veryhigh-marker', image);
+  });
+
+  mapRef.current.loadImage(poiImage, (error, image) => {
+    if (error) throw error;
+    mapRef.current.addImage('poi-marker', image);
+  });
+
+  mapRef.current.addSource('parks', {
+    type: 'geojson',
+    data: parkGeoJson,
+    cluster: true,
+    clusterMaxZoom: 14,
+    clusterRadius: 80
+  });
+
+  mapRef.current.addLayer({
+    id: 'clusters',
+    type: 'circle',
+    source: 'parks',
+    filter: ['has', 'point_count'],
+    paint: {
+      'circle-color': [
+        'step',
+        ['get', 'point_count'],
+        '#4CAF50',
+        100,
+        '#388E3C',
+        750,
+        '#2E7D32'
+      ],
+      'circle-radius': [
+        'step',
+        ['get', 'point_count'],
+        20,
+        100,
+        30,
+        750,
+        40
+      ],
+      'circle-opacity': 0.6
+    }
+  });
+
+  mapRef.current.addLayer({
+    id: 'cluster-count',
+    type: 'symbol',
+    source: 'parks',
+    filter: ['has', 'point_count'],
+    layout: {
+      'text-field': '{point_count_abbreviated}',
+      'text-font': ['DIN Offc Pro Medium', 'Arial Unicode MS Bold'],
+      'text-size': 12
+    },
+    paint: {
+      'text-color': '#FFFFFF'
+    }
+  });
+
+  mapRef.current.addLayer({
+    id: 'unclustered-point',
+    type: 'symbol',
+    source: 'parks',
+    filter: ['!', ['has', 'point_count']],
+    layout: {
+      'icon-image': 'flora-marker',
+      'icon-size': 0.5,
+      'icon-offset': [0, -0]
+    }
+  });
+
+  mapRef.current.on('click', 'unclustered-point', e => {
+    const coordinates = e.features[0].geometry.coordinates.slice();
+    const { name } = e.features[0].properties;
+
+    const popupNode = document.createElement('div');
+    ReactDOM.render(
+      <PopupContent
+        coordinates={coordinates}
+        name={name}
+        setStartCord={setStartCord}
+        setEndCord={setEndCord}
+        setWaypointAndIncrease={setWaypointAndIncrease}
+        updateStartInput={updateStartInput}
+        updateEndInput={updateEndInput}
+        updateWaypointInput={updateWaypointInput}
+        geocoderRefs={geocoderRefs}
+      />,
+      popupNode
+    );
+
+    new mapboxgl.Popup({ offset: 25, className: 'custom-popup' })
+      .setLngLat(coordinates)
+      .setDOMContent(popupNode)
+      .addTo(mapRef.current);
+  });
+
+  mapRef.current.on('mouseenter', 'clusters', () => {
+    mapRef.current.getCanvas().style.cursor = 'pointer';
+  });
+  mapRef.current.on('mouseleave', 'clusters', () => {
+    mapRef.current.getCanvas().style.cursor = '';
+  });
+
+  mapRef.current.on('mouseenter', 'unclustered-point', () => {
+    mapRef.current.getCanvas().style.cursor = 'pointer';
+  });
+  mapRef.current.on('mouseleave', 'unclustered-point', () => {
+    mapRef.current.getCanvas().style.cursor = '';
+  });
+
+  mapRef.current.on('click', 'clusters', e => {
+    const features = mapRef.current.queryRenderedFeatures(e.point, {
+      layers: ['clusters']
+    });
+    const clusterId = features[0].properties.cluster_id;
+    mapRef.current.getSource('parks').getClusterExpansionZoom(
+      clusterId,
+      (err, zoom) => {
+        if (err) return;
+
+        mapRef.current.easeTo({
+          center: features[0].geometry.coordinates,
+          zoom
+        });
+      }
+    );
+  });
 }
 
 export { animateMarkers };
